semantic-kernel-python-webapi/requirements.txt
fastapi==0.111.0
azure-identity==1.16.0
python-dotenv==1.0.1
semantic-kernel==1.32.1
